requirements.txt
fastapi
uvicorn
google-generativeai
pymupdf
python-dotenv
jinja2
